requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
black=25.1.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.12.31=h06a4308_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
detectron2=0.6=pypi_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.2=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.55.8=pypi_0
fsspec=2024.2.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
grpcio=1.70.0=pypi_0
huggingface-hub=0.27.1=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
iniconfig=2.0.0=pypi_0
iopath=0.1.9=pypi_0
jinja2=3.1.3=pypi_0
kiwisolver=1.4.8=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
mamba-ssm=2.2.4=pypi_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.0=pypi_0
mpmath=1.3.0=pypi_0
multiscaledeformableattention=1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
ninja=1.11.1.3=pypi_0
numpy=1.26.3=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.2=pypi_0
pathspec=0.12.1=pypi_0
pillow=10.2.0=pypi_0
pip=24.2=pypi_0
platformdirs=4.3.6=pypi_0
pluggy=1.5.0=pypi_0
portalocker=3.1.1=pypi_0
protobuf=5.29.3=pypi_0
pycocotools=2.0.8=pypi_0
pyparsing=3.2.1=pypi_0
pytest=8.3.4=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.5.2=pypi_0
scipy=1.15.1=pypi_0
setuptools=75.1.0=pypi_0
six=1.17.0=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.5.0=pypi_0
timm=1.0.14=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
torch=2.5.1+cu124=pypi_0
torchvision=0.20.1+cu124=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.48.1=pypi_0
triton=3.1.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
zlib=1.2.13=h5eee18b_1
